requirements.txt
flask
gunicorn
tensorflow
opencv-python
numpy
matplotlib
pillow
scikit-learn
pickle-mixin